Fix focus navigation
It should not be optional to make the current focus move when following
a link.

Note that there are bugs in browsers about how they handle this with
e.g. embedded SVG.

This is https://www.w3.org/Bugs/Public/show_bug.cgi?id=16186 (and
matches a change the wig finally made for the same bug)