Removed references to itemprop / microdata
Microdata was removed from HTML development in 2013.  The microdata
related attributes were still embedded in the spec.  This pull request
removes those.  It also changes the examples that use microdata to
examples that use RDFa Lite.  RDFa Lite is a W3C Recommendation, so is
more appropriate to use for examples.