Make stepUp and stepDown match reality for non-number/range input types.
Week, month, date, etc., perform stepping based on a rounded step value, but
this was not accounted for in the spec's algorithms.

This change adds the condition for those input types to round their value as
part of the computation of the "allowed step value". This enables the step value
to be rounded before applying the stepUp/Down optional multiplier, and also ensures
that all step values less than 1 for the affected types round-up to one.

Additionally, clarified a few things that were boggling to me when reading through
the algorithms, such as the negative step base default on the week state, and
the actual steps for working with "integral multiple" of values. These are now
re-written in terms of two new concepts "is step aligned" and the "step align"
algorithm in order to try and make this more readable and intuitive in terms of
what actions are actually taking place during the algorithm.