Remove spurious backslash characters from the spec
Bikeshed typically wants you to escape a double-backslash `[[` if you intend it to
render as-is, and not turn it into a biblio reference. However, it has some exceptions
which I discover during this work:

Turns out that if you have an element inside of the double-brackets e.g.,
[[<a>something</a>]] then you don't need to escape the leading double-brackets.

Also, any content in <pre> tags is not checked for auto-linking.