Appcache is non-conforming. (#501)
It needs to be listed in the right section.

There are still things that browsers generally need to do until it
actually disappears from the Web

Bring privacy considerations to the main section on that (as well, not
instead).

Note that a substantial amount of advice and explanation, primarily
directed at application authors, was removed. Reviewers, please check
that nothing critical went with it.