Update semantics-embedded-content.include
There is an inconsistency between 4.7.10.10 Media resources with multiple media tracks code example and loadedmetadata event specification. In HAVE_METADATA state
enough of the resource has been obtained that the duration of the resource is available. In the case of a video element, the dimensions of the video are also available. (According to this video track doesn't have to be ready yet)
According to the specification this example should rely on onaddtrack event.