Change link term
fix #919
this changes the id of the term definition, from xml-compatible-xml-compatible to xml-compatible.

As far as I can tell the spec doesn't refer to this concept. I am not sure if some other specification does, in which case we are probably better leaving old links working.

So I suggest not merging and marking #919 wontfix, but editors may think otherwise. In any event this is purely editorial in terms of conformance…