Fix #428 (#923)
This requires resource requests (for media) the end-up in the
resource fetch algorithm but are attached to a media provider
object or have a URL whose object is a media provider object
(e.g., URL.createObjectURL() from a media provider like a Blob)
to not delay the load event. These resources are treated as non
network providers and thus don't delay the load event.