editorial updates for cite attribute
editorial updates to the cite attribute as outlined in the introductory information pertaining to the `del`, `ins` and `q` elements.

Additional updates:

consistently add periods to the end of `dd`s in the `del`, `ins`, and `q` sections.

Fix a typo in the prose of `q` where there was an extra space before a comma.