update example 63 - meters
the original example used `span type=“toolbar”` which is invalid, and an `a` element with an onclick event that would likely be better displayed as a `button`.

formatted examle, updated the above mentioned elements, wrapped the code in a `section` and created a new image to reflect the changes in the code.