make lists of options lists
change the markup of the introductory example where sizes and toppings are listed to use `ul` & `li` elements instead of `div`s.

This commit also adds missing `required` attribute to the part of the introduction that mentioned adding the attribute to the pizza sizes radio buttons, as well as some minor prose source formatting.