update prose of article and section
in an attempt to better convey the purposes of each element, the prose of `article` and `section` have been updated.

Regarding `article` two new paragraphs have been added to indicate nested elements can be used within an `article` to promote logical semantics for the article’s prose. The `section` element is called out specifically to indicate how sections of an `article` may be explicitly set via this element.

Regarding the `section` element, two of the existing notes have been updated to further callout the differences between a `section` and `article`’s definitions.