revise prose for `link[rel]`
previously, the prose for `link[rel]` had some ambigutity in whether or not a `rel` attribute was necessary.  e.g. wording like “if present” vs a later line reading “must have” concerning the existence of the attribute.

This revision consolidates prose pertaining to the existence of the attribute, and up front states a `link` must have a `rel` attribute.