1335 rel rev cleanup (#1341)
* source cleanup

no changes to prose here. merely adhering to the 100char line limit

* revise prose for `link[rel]`

previously, the prose for `link[rel]` had some ambigutity in whether or not a `rel` attribute was necessary.  e.g. wording like “if present” vs a later line reading “must have” concerning the existence of the attribute.

This revision consolidates prose pertaining to the existence of the attribute, and up front states a `link` must have a `rel` attribute.

* add mention of `link`s to heading

while not porting over all the information specific to the `link` element, these commits aim to add more references to the element to mitigate the fact it’s called out in some paragraphs, but was not overtly mentioned in the heading or in other applicable instances.