update prose for table border attribute
fixes #1345

editorial update to the border prose to indicate what happens when a value of `border=“foo”` is added to a table, and provide some additional explination on what the attribute should indicate to user agents.

add prose that authors should not use this attribute for styling purposes, and should use CSS instead.