forms introduction prose fix
the ordering of the `autocomplete` step in the forms introduction may have been modified due to a faulty merge, as it was appearing in the wrong order.

This commit fixes the ordering of the form introduction to place the `autocomplet` example prior to the radio buttons prose, where it belongs.