Remove <applet>.
This leaves the parser logic intact which is aligned with upstream, but removes
all support, handling, and rendering rules for applet elements.

There is one example in the embedded content section that still references
applets, which was tweaked to now serve as an example of how to provide
fallback content in the context of removed features.

* Upstream: whatwg/html@b9b9d60f9451c05f8e8d542e4094f987ae1a5fb8