address unclear figcaption usage
Fixes #1540

Addresses point to clarif content model for `figure` element.

Mention in the prose of the `figcaption` section that there should be no more than one child `figcaption`  to a single parent `figure`.

Additionally pull over prose from `table` section, about how a `figcaption` should be used instead of a `caption` if a `table` is the primary content of a `figure`.

Further, added new example showing how 2 `figcaption` elements should be avoided and instead coded as a single `figcaption` to a `figure` showcasing a code snippet.

Includes other file cleanup.